requirements.txt
addict==2.4.0
beautifulsoup4==4.12.2
numpy==1.25.0
opencv_python==4.8.0.74
pandas==2.0.3
PyYAML==6.0
scikit_learn==1.3.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
